separating dashboard skeleton
